move rta reconnect logic into separate file
